# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "Phase 1 Swagger Integration"
  version: "1.0.0"
  title: "iProject"
host: "localhost:3002"
...
schemes:
- "http"
paths:
...
  /api/getToken:
      post:
        tags:
        - "public"
        summary: "Signup User"
        description: ""
        operationId: "getToken"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - name: "body"
          in: "body"
          description: "Reset object"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
        responses:
          200:
            description: "successful operation"
            schema:
              $ref: "#/definitions/GetTokenResponse"
          400:
            description: "multiple reasons"
            schema:
              type: array
              items:
                type: "string"
                enum:
                - "Invalid User"
                - "User is already verified"
                - "Link has been expired. Please check recovery option by clicking on forgot password"
  /auth/info:
    get:
      tags:
      - "user"
      summary: "Fetch User"
      description: "Fetch User"
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      security:
      - api_key: []
  /auth/user:
    post:
      tags:
      - "user"
      summary: "Signup User"
      description: "Signup User"
      operationId: "createUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create new user"
        required: true
        schema:
          $ref: "#/definitions/createUser"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SignupResponse"
        422:
          description: "Email and permission are required."
        400:
          description: "multiple reasons"
          schema:
            type: array
            items:
              type: "string"
              enum:
              - "Username must be less than 10 characters."
              - "The email is already registered."
              - "Email is too long, please use shorter email address"
              - "Email must be in valid format"
      security:
      - api_key: []
  /auth/user/{userId}:
    get:
      tags:
      - "user"
      summary: "Fetch User Details By Id"
      description: "Fetch User Details"
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      security:
      - api_key: []
    delete:
      tags:
      - "user"
      summary: "Removing User Details"
      description: "Removing User Details"
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "userId"
        in: "path"
        description: "User id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
      security:
      - api_key: []
    put:
      tags:
      - "user"
      summary: "Updating User Details"
      description: "Updating User Details"
      operationId: "updateProfileByID"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "Id that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/updateUser"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      security:
      - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "uuid"
      userName:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      isVerified:
        type: "boolean"
      permissionId:
        type: "integer"
  updateUser:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
      userName:
        type: "string"
      isVerified:
        type: "boolean"
  createUser:
    type: "object"
    properties:
      firstName:
        type: "string"
      lastName:
        type: "string"
      userName:
        type: "string"
      email:
        type: "string"
      updatedBy:
        type: "integer"
      companyId:
        type: "integer"
      permissionId:
        type: "integer"
...
  SignupResponse:
    type: "object"
    properties:
      userName:
        type: "string"
      email:
        type: "string"
...
  GetTokenResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
      message:
        type: "string"
      token:
        type: "string"
...
  ApiResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
      message:
        type: "string"